[css-grid] Don't resolve Default Alignment 'normal 'values
Our logic to detect style changes that may affect the grid item's size
checks whether the value has changed from or to stretch. In case of any
of the involved values is 'normal', we are using the grid's 'normal'
behavior, which as per specs it'll be resolved as 'stretch'.

However, in case of replaced items we must use resolve any 'normal'
value as 'start', instead of 'stretch', so it uses its intrinsic size.

Hence, we can't assume Default Alignment 'normal' values can be resolved
to 'stretch' in all the cases; we will need to evaluate each grid item's
nature (obviously if its Self Alignment value is 'auto').

Bug: 750681
Change-Id: I869848904c75402b6ab247509090a86bac4030a1
Reviewed-on: https://chromium-review.googlesource.com/611802
Reviewed-by: Manuel Rego Casasnovas <rego@igalia.com>
WPT-Export-Revision: 5200f29a2c72c304d8ef111ecf174242008fdf84