Hide relatively positioned layers in collapsed columns
Add a bit to PaintLayer so that when the layer is in a collapsed cell, it
is not painted. Add a bit to LayoutState so that children of collapsed
cells know that they have a collapsed cell as an ancestor.

Bug:
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2
Change-Id: I865bb06f8b879e180d4aa6de57010e2609afb8b1
Reviewed-on: https://chromium-review.googlesource.com/644173
WPT-Export-Revision: 97d02eab072602a0b242d3199bbada1b041f0999